requirements.txt
aiogram==2.25.2
python-dotenv==1.0.0
aiofiles==23.2.1